Allow horizontal axis to be ordered by call site
This adds the option --call-site to stackcollapse-perf.pl, which will
prefix each function name with its call site. This has two effects:
functions at the same depth are no longer sorted alphabetically, but
rather according to where in the parent function they are called. It
also distinguishes when the same function is called from different call
sites. In flamegraph.pl, the call site prefixes are removed from the
output.