display Left/Right/Difference samples in differential flamegraphs
When calling flamegraph.pl in differential mode with:

$ difffolded.pl -n Left.perf-folded Right.perf-folded | flamegraph.pl > out.svg

flamegraph.pl used to only display the total number of cycles from the Left
diffed file.  With this change, flamegraph.pl also shows the number of samples
collected only in the current function, i.e., without accounting for the called
functions, for Left, Right, and the difference Left - Right.