Handle SIGINT
When FlameGraph is piped to another tool, pressing Ctrl+C causes the 
loss of all the input produced so far. The proposed change handles the 
SIGINT and interrups FlameGraph only on a second press of Ctrl+C. This 
allows any input received to be processed and still produce a partial 
graph.