"Simplify" regular expressions
Handles the given TODO:

  # XXX todo: redo these REs as one. It's repeated below.

By converting regexp calls to a single parsing call per line.

This has two benefits:

- Simplifies if logic
- Has about a 2x speed up when generating the graph

This is accomplished by converting the regexp to instead "work
backwards" and capture data from the end of the line instead of the
front.  Since the beginning of the line has to be very greedy since it
can include numbers and spaces itself, it is "simpler" to write regexp
in a way that parses the data from the back by looking for 1 to 2
digits at the end of the line.  This way, we can be greedy with those
captures in one single expression instead of requiring that separate
match calls are done to see if we have another trailing digit with the
second call regexp match.

As a result, there is nearly a 2x speed up with the data I was testing
against, and the graphs seem to produce the same data each time.

Before
------

(fastest of 5 runs on a 13M *.stackcollapse file)

real    0m1.308s
user    0m1.268s
sys     0m0.034s

After
------

(fastest of 5 runs on a 13M *.stackcollapse file)

real    0m0.798s
user    0m0.761s
sys     0m0.032s