Assign delta to each frame but not only the top one
delta should be the difference of num_d1 and num_d2 for the same frame,
but the current calculation in flow() only computes the difference when
the frame shows up at the first time, and discards the other differences
when the frame shows up afterwards. Thus change the code to calculate
all differences as long as the frame shows up by traversing $this frame
stacks.

Signed-off-by: Fei Li <lifei.shirley@bytedance.com>