Fix the diff percentage and color for diff-flamegraph
In the current code, the $deltapct% means the diff(d2 - d1) / sum(d2),
which is not appropriate to exactly show the difference. Change this
value to the difference of percentage of each sampling.

Give an example, supposing
- d1: do_futex 300/5000 = 6%
- d2: do_futex 90/3000 = 3%
- diff result:
  - Before change: do_futex (90 samples, 3%; -7%) //(90-300)/3000=-7%
  - After change: do_futex (90 samples, 3%; -3%) //more accurate

Besides, to show the difference of color, use $deltapct/$maxdeltapct
instead of previous $delta/$maxdelta.

Signed-off-by: Fei Li <lifei.shirley@bytedance.com>