[stackcollapse] Improve the stackcollapse-tracecmd.pl to apply on Ftrace too
As the stackcollparse-tracecmd.pl is originally written for the
trace-cmd output, however, the trace-cmd is from the Ftrace output. So,
the patch is to improve the parsing of the format to apply on the Ftrace
original output too(/sys/kerne/debug/tracing/trace).

In the future, the parser can be improved in many perspectives:
1). memory: To base on the byte size/number of pages allocated/latency/order/pid/migrate type...
2). IO: to based on the latency.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>