[stackcollapse] Add the size based memory allocation flamegraph
When using Ftrace/trace-cmd to profile the memory allocation source, it
will be much more useful to generate the memory allocation size
flamegraph instead of the sampling count flamegraph.

The steps:
1). sudo trace-cmd record -l __alloc_pages_nodemask -e kmem:mm_page_alloc  -T sleep 5
2). trace-cmd report --cpu 1 > trace-cmd.log # generate the event only
from CPU 1
3). FlameGraph/stackcollapse-tracecmd.pl --process_name --page_type --page_order  ./trace-cmd.log | ~/os/FlameGraph/flamegraph.pl --countname kb > ./trace-cmd.svg
4). firefox ./trace-cmd.svg # or google-chrome-stable

The --page_type is used to classify based on the migration/gfp_flags.
The --page_order is to classify the callstack by the allcation order.

The stackcollapse-tracecmd.pl supports both ftrace/trace-cmd format
parsing.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>