Fall back to symbol+offset when calculating source lines if mod+pc fails
When using perf to profile jni code in java with help of a map agent [1],
the output of perf script contain raw $pc values for dynamically loaded
libraries. This does not get translated into source lines when using the
'stackcollapse-perf.pl --source --inline' command.
But perf script also produces symbol+offset in outputs, so this commit
adds a fallback that, if normal means fail, calculates address in mod by
getting the symbol address with nm and adding to the offset, and retries
with source line calculation for this address, producing a valid result.

[1] https://github.com/jvm-profiling-tools/perf-map-agent