Add stackcollapse-trace-cmd.rb
This is a stackcollapse processor for the output of `trace-cmd`,
specifically the output of `trace-cmd report` when processing a
recording with the `function_graph` plugin.

It is designed to be used with flamegraph.pl --flamechart, and
spacers are inserted to prevent individual calls from being merged.

There is also a latency cutoff parameter that can be used to filter
for slow calls.